requirements.txt
kafka-python==2.0.2
pandas==2.0.3
scikit-learn==1.3.0
mysql-connector-python==8.1.0
joblib==1.3.2
numpy==1.24.3